video: Moved VideoBackend from core to video
