+++ core/src/backend/video/software/mod.rs
@@ -1,9 +1,9 @@
 //! Pure software video decoding backend.
 
-use crate::backend::video::VideoBackend;
 use generational_arena::Arena;
 use ruffle_render::backend::RenderBackend;
 use ruffle_render::bitmap::{Bitmap, BitmapFormat, BitmapHandle, BitmapInfo};
+use ruffle_video::backend::VideoBackend;
 use ruffle_video::error::Error;
 use ruffle_video::frame::{DecodedFrame, EncodedFrame, FrameDependency};
 use ruffle_video::VideoStreamHandle;
