+++ video/src/lib.rs
@@ -1,5 +1,6 @@
 use generational_arena::Index;
 
+pub mod backend;
 pub mod error;
 pub mod frame;
